Update pytz and tzlocal to resolve issue with Brazil DST with timezone
As Brazil does not have Daylight Saving Time anymore as decided by it's government, pytz needed to be updated to have the new data about the country timezones.
And as some in pytz calls in bazarr used tzlocal functions, that one needed also to be updated.
Works fine now, tested on Brazil - Timezone America/Sao_Paulo

This commit fixes #641