Start documenting and tracking direct dependency versions
First list of dependencies and their versions as documented in
libs/version.txt. The versions have been checked by installing them
in a temporary directory and running the diff command on them and the
ones found in the git repository of bazarr.

There are more direct dependencies that are not listed here yet.
Those have to be checked manually and the correct versions should be
added to this file.

And then there's a group of direct dependencies that have been bundled,
and then modified without the modifications being upstreamed.

Will have to figure out how to handle those.
But first steps are made to make this administration a bit easier.

To install and bundle these dependencies (including their dependencies)
one can use the following pip command from the root of this porject:

PIP_TARGET=./libs/ pip install -r ./libs/requirements.txt --upgrade