Try to avoid guessit internal exception (#1858)
I was unable to reproduce the exception, but I found a strange pattern:
whenever the hints["expected_title"] value had the same value as the
title found in the filename, there was a list from the rebulk module
constantly growing. Once I remove the expected_title "hint", this
behavior is gone.

This change is transparent. If for some reason guessit returns a
different movie/series title, Bazarr will refine it with the database
anyway.