no log: Made 'id' description clearer (#2209)
I've been playing with the API for the last few hours and I noticed that this wasn't clear that the ID also needs to be set if it's a movie. At first I gave it a -1 ID because I presumed the ID wasn't needed as it wasn't an episode. 

This small change should help people in the future when utilizing the API.