Added Sync Only Monitored Episodes feature
Implemented algorithm. that (when the switch is on) will only check episodes that are monitored, or whose monitored state does not agree between Bazarr and Sonarr.
Added statistics to indicate effectiveness of algorithm and logging to help with any potential further debugging.
Added trace() method to wrap logging.debug() calls with a feature prefix to make it easy to filter out these messages when Bazarr is in logging DEBUG mode.