Workaround for pop-up menu being clipped
This adds the minimum amount of space needed to the parent form to ensure that the entire pop-up menu is displayed.
The reason it's a workaround is because the space is always added, even if not needed. If I knew how to do math inside a .tsx file, I would have done so to adjust the amount added, if that is even possible.
An even better solution would be to somehow attach the menu to the main window, instead of the modal dialog. However, after much searching on this and trying several things, I couldn't make it happen. Perhaps someone with Mantine experience can do better. In the meantime, this works.
This should fix issue #2348 which actually happens for both TV Series and Movies.