Fix best forced subtitle downloads
Searching for the best forced subtitles for a given language was
resulting in all forced subtitles for that language being downloaded in
descending score order until the minimum score was reached.

Not only did this burn through any download limits imposed by providers,
it left poor quality subtitles downloaded (which could later be
automatically upgraded to the first choice).

This change uses the string conversion of Language objects instead of
their basenames when working out when to stop downloading subtitles, as
this takes into account the forced flag while still ignoring the hearing
impaired flag.