Handle multiline log output without proper log file timestamp headers
Sometimes, we will get errors in the log file of the form: "ffprobe cannot analyze this video file...Could it be corrupted?"
These are generated by KnowIt 0.5.3 and spew many lines of information, including stack traces, but in simple text, not in proper log file format.
This fix captures all of that output for display under the exception column in the log file viewer page.