Revert to using reqparse
This allows complete backwards compatability with userscripts
targeting the previous implementation at the cost of accurate
swagger documentation. It seems like this might be the best case
scenario due to what seem to be some flask-restx limitations
around accepting headers when using the expect decorater with
models. Unfortunately it means we need to handle a lot more
validation inside the function itself since we can't rely on
flask-restx.